Beef up the run-tests.sh script
The script will now refuse to run on an existing X server and instead
spawn its own instance to make sure the X server runs at the same color
depth as the GLES test program.

The new --disable-sync option can be used to disable synchronization of
rendering operations with the VBLANK. A --performance option is added
which will automatically set the CPU governor to "performance", causing
the CPU to always operate at the maximum frequency. Finally a new mode
can be enabled with the --regenerate option, which causes the test
pattern to be regenerated for each frame in order to simulate dynamic
input data.

Signed-off-by: Thierry Reding <thierry.reding@avionic-design.de>