Major rework
This allows us to get rid of the GLib/GIO dependencies and generally
removes any remnants of the GStreamer plugin ancestry.

In order to make it easier to test various combinations of test-cases, a
dynamic pipeline is introduced which can be created by specifying a list
of stages on the command-line.

Signed-off-by: Thierry Reding <thierry.reding@avionic-design.de>