Display megatexels per second in summary
Since the textures are mapped 1:1 to screen space, the number of texels
processed is equal to the number of pixels displayed. The value shown is
the number of megatexels that are processed through the given pipeline
and not the absolute throughput (which would be dependent on the number
of stages in the pipeline).

Signed-off-by: Thierry Reding <thierry.reding@avionic-design.de>