Wait for X server to shutdown
When shutting down the X server at the end of a test series, wait until
it has stopped by watching for /tmp/.X0-lock to disappear. Otherwise it
can happen that we try to run another series of tests, which will fail
because it detects that another X server is still running.

Signed-off-by: Thierry Reding <thierry.reding@avionic-design.de>